+++ node/src/GlideClient.ts
@@ -6,22 +6,16 @@ import * as net from "net";
 import {
     BaseClient,
     BaseClientConfiguration,
+    convertGlideRecordToRecord,
     Decoder,
     DecoderOption,
+    GlideRecord,
+    GlideReturnType,
     GlideString,
     PubSubMsg,
     ReadFrom, // eslint-disable-line @typescript-eslint/no-unused-vars
-    ReturnType,
 } from "./BaseClient";
 import {
-    FlushMode,
-    FunctionListOptions,
-    FunctionListResponse,
-    FunctionRestorePolicy,
-    FunctionStatsFullResponse,
-    InfoOptions,
-    LolwutOptions,
-    SortOptions,
     createClientGetName,
     createClientId,
     createConfigGet,
@@ -57,6 +51,14 @@ import {
     createSortReadOnly,
     createTime,
     createUnWatch,
+    FlushMode,
+    FunctionListOptions,
+    FunctionListResponse,
+    FunctionRestorePolicy,
+    FunctionStatsFullResponse,
+    InfoOptions,
+    LolwutOptions,
+    SortOptions,
 } from "./Commands";
 import { connection_request } from "./ProtobufMessage";
 import { Transaction } from "./Transaction";
@@ -79,7 +81,7 @@ export namespace GlideClientConfiguration {
         Pattern = 1,
     }
 
-    export type PubSubSubscriptions = {
+    export interface PubSubSubscriptions {
         /**
          * Channels and patterns by modes.
          */
@@ -96,7 +98,7 @@ export namespace GlideClientConfiguration {
          */
         /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
         context?: any;
-    };
+    }
 }
 
 export type GlideClientConfiguration = BaseClientConfiguration & {
@@ -191,8 +193,8 @@ export class GlideClient extends BaseClient {
     public async exec(
         transaction: Transaction,
         options?: DecoderOption,
-    ): Promise<ReturnType[] | null> {
-        return this.createWritePromise<ReturnType[] | null>(
+    ): Promise<GlideReturnType[] | null> {
+        return this.createWritePromise<GlideReturnType[] | null>(
             transaction.commands,
             options,
         ).then((result) =>
@@ -224,7 +226,7 @@ export class GlideClient extends BaseClient {
     public async customCommand(
         args: GlideString[],
         options?: DecoderOption,
-    ): Promise<ReturnType> {
+    ): Promise<GlideReturnType> {
         return this.createWritePromise(createCustomCommand(args), options);
     }
 
@@ -397,7 +399,10 @@ export class GlideClient extends BaseClient {
         parameters: string[],
         options?: DecoderOption,
     ): Promise<Record<string, GlideString>> {
-        return this.createWritePromise(createConfigGet(parameters), options);
+        return this.createWritePromise<GlideRecord<GlideString>>(
+            createConfigGet(parameters),
+            options,
+        ).then(convertGlideRecordToRecord);
     }
 
     /**
@@ -649,7 +654,13 @@ export class GlideClient extends BaseClient {
     public async functionList(
         options?: FunctionListOptions & DecoderOption,
     ): Promise<FunctionListResponse> {
-        return this.createWritePromise(createFunctionList(options), options);
+        return this.createWritePromise<GlideRecord<unknown>[]>(
+            createFunctionList(options),
+            options,
+        ).then(
+            (res) =>
+                res.map(convertGlideRecordToRecord) as FunctionListResponse,
+        );
     }
 
     /**
@@ -700,7 +711,13 @@ export class GlideClient extends BaseClient {
     public async functionStats(
         options?: DecoderOption,
     ): Promise<FunctionStatsFullResponse> {
-        return this.createWritePromise(createFunctionStats(), options);
+        return this.createWritePromise<GlideRecord<unknown>>(
+            createFunctionStats(),
+            options,
+        ).then(
+            (res) =>
+                convertGlideRecordToRecord(res) as FunctionStatsFullResponse,
+        );
     }
 
     /**