+++ node/src/GlideClusterClient.ts
@@ -8,10 +8,12 @@ import {
     BaseClientConfiguration,
     Decoder,
     DecoderOption,
+    GlideRecord,
     GlideString,
     PubSubMsg,
     ReadFrom, // eslint-disable-line @typescript-eslint/no-unused-vars
-    ReturnType,
+    GlideReturnType,
+    convertGlideRecordToRecord,
 } from "./BaseClient";
 import {
     FlushMode,
@@ -64,23 +66,23 @@ import { connection_request } from "./ProtobufMessage";
 import { ClusterTransaction } from "./Transaction";
 
 /** An extension to command option types with {@link Routes}. */
-export type RouteOption = {
+export interface RouteOption {
     /**
      * Specifies the routing configuration for the command.
      * The client will route the command to the nodes defined by `route`.
      */
     route?: Routes;
-};
+}
 
 /**
  * Represents a manually configured interval for periodic checks.
  */
-export type PeriodicChecksManualInterval = {
+export interface PeriodicChecksManualInterval {
     /**
      * The duration in seconds for the interval between periodic checks.
      */
     duration_in_sec: number;
-};
+}
 
 /**
  * Periodic checks configuration.
@@ -122,7 +124,7 @@ export namespace GlideClusterClientConfiguration {
         Sharded = 2,
     }
 
-    export type PubSubSubscriptions = {
+    export interface PubSubSubscriptions {
         /**
          * Channels and patterns by modes.
          */
@@ -139,7 +141,7 @@ export namespace GlideClusterClientConfiguration {
          */
         /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
         context?: any;
-    };
+    }
 }
 export type GlideClusterClientConfiguration = BaseClientConfiguration & {
     /**
@@ -163,7 +165,39 @@ export type GlideClusterClientConfiguration = BaseClientConfiguration & {
  */
 export type ClusterResponse<T> = T | Record<string, T>;
 
-export type SlotIdTypes = {
+/**
+ * @internal
+ * Type which returns GLIDE core for commands routed to multiple nodes.
+ * Should be converted to {@link ClusterResponse}.
+ */
+type ClusterGlideRecord<T> = GlideRecord<T> | T;
+
+/**
+ * @internal
+ * Convert {@link ClusterGlideRecord} to {@link ClusterResponse}.
+ *
+ * @param res - Value received from Glide core.
+ * @param isRoutedToSingleNodeByDefault - Default routing policy.
+ * @param route - The route.
+ * @returns Converted value.
+ */
+function convertClusterGlideRecord<T>(
+    res: ClusterGlideRecord<T>,
+    isRoutedToSingleNodeByDefault: boolean,
+    route?: Routes,
+): ClusterResponse<T> {
+    const isSingleNodeResponse =
+        // route not given and command is routed by default to a random node
+        (!route && isRoutedToSingleNodeByDefault) ||
+        // or route is given and it is a single node route
+        (Boolean(route) && route !== "allPrimaries" && route !== "allNodes");
+
+    return isSingleNodeResponse
+        ? (res as T)
+        : convertGlideRecordToRecord(res as GlideRecord<T>);
+}
+
+export interface SlotIdTypes {
     /**
      * `replicaSlotId` overrides the `readFrom` configuration. If it's used the request
      * will be routed to a replica, even if the strategy is `alwaysFromPrimary`.
@@ -174,9 +208,9 @@ export type SlotIdTypes = {
      * Unless the slot is known, it's better to route using `SlotKeyTypes`
      */
     id: number;
-};
+}
 
-export type SlotKeyTypes = {
+export interface SlotKeyTypes {
     /**
      * `replicaSlotKey` overrides the `readFrom` configuration. If it's used the request
      * will be routed to a replica, even if the strategy is `alwaysFromPrimary`.
@@ -186,10 +220,10 @@ export type SlotKeyTypes = {
      * The request will be sent to nodes managing this key.
      */
     key: string;
-};
+}
 
 /// Route command to specific node.
-export type RouteByAddress = {
+export interface RouteByAddress {
     type: "routeByAddress";
     /**
      *The endpoint of the node. If `port` is not provided, should be in the `${address}:${port}` format, where `address` is the preferred endpoint as shown in the output of the `CLUSTER SLOTS` command.
@@ -199,7 +233,7 @@ export type RouteByAddress = {
      * The port to access on the node. If port is not provided, `host` is assumed to be in the format `${address}:${port}`.
      */
     port?: number;
-};
+}
 
 export type Routes =
     | SingleNodeRoute
@@ -306,7 +340,7 @@ export class GlideClusterClient extends BaseClient {
     public async customCommand(
         args: GlideString[],
         options?: RouteOption & DecoderOption,
-    ): Promise<ClusterResponse<ReturnType>> {
+    ): Promise<ClusterResponse<GlideReturnType>> {
         const command = createCustomCommand(args);
         return super.createWritePromise(command, options);
     }
@@ -333,8 +367,8 @@ export class GlideClusterClient extends BaseClient {
         options?: {
             route?: SingleNodeRoute;
         } & DecoderOption,
-    ): Promise<ReturnType[] | null> {
-        return this.createWritePromise<ReturnType[] | null>(
+    ): Promise<GlideReturnType[] | null> {
+        return this.createWritePromise<GlideReturnType[] | null>(
             transaction.commands,
             options,
         ).then((result) =>
@@ -401,10 +435,10 @@ export class GlideClusterClient extends BaseClient {
     public async info(
         options?: { sections?: InfoOptions[] } & RouteOption,
     ): Promise<ClusterResponse<string>> {
-        return this.createWritePromise<ClusterResponse<string>>(
+        return this.createWritePromise<ClusterGlideRecord<string>>(
             createInfo(options?.sections),
             { decoder: Decoder.String, ...options },
-        );
+        ).then((res) => convertClusterGlideRecord(res, false, options?.route));
     }
 
     /**
@@ -437,10 +471,10 @@ export class GlideClusterClient extends BaseClient {
     public async clientGetName(
         options?: RouteOption & DecoderOption,
     ): Promise<ClusterResponse<GlideString | null>> {
-        return this.createWritePromise<ClusterResponse<GlideString | null>>(
+        return this.createWritePromise<ClusterGlideRecord<GlideString | null>>(
             createClientGetName(),
             options,
-        );
+        ).then((res) => convertClusterGlideRecord(res, true, options?.route));
     }
 
     /**
@@ -511,10 +545,10 @@ export class GlideClusterClient extends BaseClient {
     public async clientId(
         options?: RouteOption,
     ): Promise<ClusterResponse<number>> {
-        return this.createWritePromise<ClusterResponse<number>>(
+        return this.createWritePromise<ClusterGlideRecord<number>>(
             createClientId(),
             options,
-        );
+        ).then((res) => convertClusterGlideRecord(res, true, options?.route));
     }
 
     /**
@@ -548,7 +582,11 @@ export class GlideClusterClient extends BaseClient {
         parameters: string[],
         options?: RouteOption & DecoderOption,
     ): Promise<ClusterResponse<Record<string, GlideString>>> {
-        return this.createWritePromise(createConfigGet(parameters), options);
+        return this.createWritePromise<
+            ClusterGlideRecord<GlideRecord<GlideString>>
+        >(createConfigGet(parameters), options).then((res) =>
+            convertGlideRecordToRecord(res as GlideRecord<string>),
+        );
     }
 
     /**
@@ -608,7 +646,10 @@ export class GlideClusterClient extends BaseClient {
         message: GlideString,
         options?: RouteOption & DecoderOption,
     ): Promise<ClusterResponse<GlideString>> {
-        return this.createWritePromise(createEcho(message), options);
+        return this.createWritePromise<ClusterGlideRecord<GlideString>>(
+            createEcho(message),
+            options,
+        ).then((res) => convertClusterGlideRecord(res, true, options?.route));
     }
 
     /**
@@ -644,10 +685,10 @@ export class GlideClusterClient extends BaseClient {
     public async time(
         options?: RouteOption,
     ): Promise<ClusterResponse<[string, string]>> {
-        return this.createWritePromise(createTime(), {
-            decoder: Decoder.String,
-            ...options,
-        });
+        return this.createWritePromise<ClusterGlideRecord<[string, string]>>(
+            createTime(),
+            options,
+        ).then((res) => convertClusterGlideRecord(res, true, options?.route));
     }
 
     /**
@@ -699,14 +740,16 @@ export class GlideClusterClient extends BaseClient {
     public async lolwut(
         options?: LolwutOptions & RouteOption,
     ): Promise<ClusterResponse<string>> {
-        return this.createWritePromise(createLolwut(options), {
-            decoder: Decoder.String,
-            ...options,
-        });
+        return this.createWritePromise<ClusterGlideRecord<string>>(
+            createLolwut(options),
+            options,
+        ).then((res) => convertClusterGlideRecord(res, true, options?.route));
     }
 
     /**
      * Invokes a previously loaded function.
+     *
+     * The command will be routed to a random node, unless `route` is provided.
      *
      * @see {@link https://valkey.io/commands/fcall/|valkey.io} for details.
      * @remarks Since Valkey version 7.0.0.
@@ -726,12 +769,17 @@ export class GlideClusterClient extends BaseClient {
         func: GlideString,
         args: GlideString[],
         options?: RouteOption & DecoderOption,
-    ): Promise<ClusterResponse<ReturnType>> {
-        return this.createWritePromise(createFCall(func, [], args), options);
+    ): Promise<ClusterResponse<GlideReturnType>> {
+        return this.createWritePromise<ClusterGlideRecord<GlideReturnType>>(
+            createFCall(func, [], args),
+            options,
+        ).then((res) => convertClusterGlideRecord(res, true, options?.route));
     }
 
     /**
      * Invokes a previously loaded read-only function.
+     *
+     * The command will be routed to a random node, unless `route` is provided.
      *
      * @see {@link https://valkey.io/commands/fcall/|valkey.io} for details.
      * @remarks Since Valkey version 7.0.0.
@@ -752,11 +800,11 @@ export class GlideClusterClient extends BaseClient {
         func: GlideString,
         args: GlideString[],
         options?: RouteOption & DecoderOption,
-    ): Promise<ClusterResponse<ReturnType>> {
-        return this.createWritePromise(
+    ): Promise<ClusterResponse<GlideReturnType>> {
+        return this.createWritePromise<ClusterGlideRecord<GlideReturnType>>(
             createFCallReadOnly(func, [], args),
             options,
-        );
+        ).then((res) => convertClusterGlideRecord(res, true, options?.route));
     }
 
     /**
@@ -851,6 +899,8 @@ export class GlideClusterClient extends BaseClient {
     /**
      * Returns information about the functions and libraries.
      *
+     * The command will be routed to a random node, unless `route` is provided.
+     *
      * @see {@link https://valkey.io/commands/function-list/|valkey.io} for details.
      * @remarks Since Valkey version 7.0.0.
      *
@@ -879,12 +929,28 @@ export class GlideClusterClient extends BaseClient {
     public async functionList(
         options?: FunctionListOptions & DecoderOption & RouteOption,
     ): Promise<ClusterResponse<FunctionListResponse>> {
-        return this.createWritePromise(createFunctionList(options), options);
+        return this.createWritePromise<
+            GlideRecord<unknown> | GlideRecord<unknown>[]
+        >(createFunctionList(options), options).then((res) =>
+            res.length == 0
+                ? (res as FunctionListResponse) // no libs
+                : ((Array.isArray(res[0])
+                      ? // single node response
+                        ((res as GlideRecord<unknown>[]).map(
+                            convertGlideRecordToRecord,
+                        ) as FunctionListResponse)
+                      : // multi node response
+                        convertGlideRecordToRecord(
+                            res as GlideRecord<unknown>,
+                        )) as ClusterResponse<FunctionListResponse>),
+        );
     }
 
     /**
      * Returns information about the function that's currently running and information about the
      * available execution engines.
+     *
+     * The command will be routed to all primary nodes, unless `route` is provided.
      *
      * @see {@link https://valkey.io/commands/function-stats/|valkey.io} for details.
      * @remarks Since Valkey version 7.0.0.
@@ -932,7 +998,14 @@ export class GlideClusterClient extends BaseClient {
     public async functionStats(
         options?: RouteOption & DecoderOption,
     ): Promise<ClusterResponse<FunctionStatsSingleResponse>> {
-        return this.createWritePromise(createFunctionStats(), options);
+        return this.createWritePromise<
+            ClusterGlideRecord<GlideRecord<unknown>>
+        >(createFunctionStats(), options).then(
+            (res) =>
+                convertGlideRecordToRecord(
+                    res,
+                ) as ClusterResponse<FunctionStatsSingleResponse>,
+        );
     }
 
     /**
@@ -975,10 +1048,10 @@ export class GlideClusterClient extends BaseClient {
     public async functionDump(
         options?: RouteOption,
     ): Promise<ClusterResponse<Buffer>> {
-        return this.createWritePromise(createFunctionDump(), {
-            decoder: Decoder.Bytes,
-            ...options,
-        });
+        return this.createWritePromise<ClusterGlideRecord<Buffer>>(
+            createFunctionDump(),
+            { decoder: Decoder.Bytes, ...options },
+        ).then((res) => convertClusterGlideRecord(res, true, options?.route));
     }
 
     /**
@@ -1084,10 +1157,8 @@ export class GlideClusterClient extends BaseClient {
      * console.log("Number of keys across all primary nodes: ", numKeys);
      * ```
      */
-    public async dbsize(
-        options?: RouteOption,
-    ): Promise<ClusterResponse<number>> {
-        return this.createWritePromise(createDBSize(), options);
+    public async dbsize(options?: RouteOption): Promise<number> {
+        return this.createWritePromise<number>(createDBSize(), options);
     }
 
     /** Publish a message on pubsub channel.
@@ -1119,7 +1190,7 @@ export class GlideClusterClient extends BaseClient {
     public async publish(
         message: GlideString,
         channel: GlideString,
-        sharded: boolean = false,
+        sharded = false,
     ): Promise<number> {
         return this.createWritePromise(
             createPublish(message, channel, sharded),
@@ -1162,28 +1233,35 @@ export class GlideClusterClient extends BaseClient {
     /**
      * Returns the number of subscribers (exclusive of clients subscribed to patterns) for the specified shard channels.
      *
-     * Note that it is valid to call this command without channels. In this case, it will just return an empty map.
-     * The command is routed to all nodes, and aggregates the response to a single map of the channels and their number of subscriptions.
-     *
      * @see {@link https://valkey.io/commands/pubsub-shardnumsub/|valkey.io} for details.
+     * @remarks The command is routed to all nodes, and aggregates the response into a single list.
      *
      * @param channels - The list of shard channels to query for the number of subscribers.
-     *                   If not provided, returns an empty map.
-     * @returns A map where keys are the shard channel names and values are the number of subscribers.
+     * @param options - (Optional) see {@link DecoderOption}.
+     * @returns A list of the shard channel names and their numbers of subscribers.
      *
      * @example
      * ```typescript
      * const result1 = await client.pubsubShardnumsub(["channel1", "channel2"]);
-     * console.log(result1); // Output: { "channel1": 3, "channel2": 5 }
+     * console.log(result1); // Output:
+     * // [{ channel: "channel1", numSub: 3}, { channel: "channel2", numSub: 5 }]
      *
-     * const result2 = await client.pubsubShardnumsub();
-     * console.log(result2); // Output: {}
+     * const result2 = await client.pubsubShardnumsub([]);
+     * console.log(result2); // Output: []
      * ```
      */
     public async pubsubShardNumSub(
-        channels?: string[],
-    ): Promise<Record<string, number>> {
-        return this.createWritePromise(createPubSubShardNumSub(channels));
+        channels: GlideString[],
+        options?: DecoderOption,
+    ): Promise<{ channel: GlideString; numSub: number }[]> {
+        return this.createWritePromise<GlideRecord<number>>(
+            createPubSubShardNumSub(channels),
+            options,
+        ).then((res) =>
+            res.map((r) => {
+                return { channel: r.key, numSub: r.value };
+            }),
+        );
     }
 
     /**
@@ -1298,7 +1376,10 @@ export class GlideClusterClient extends BaseClient {
     public async lastsave(
         options?: RouteOption,
     ): Promise<ClusterResponse<number>> {
-        return this.createWritePromise(createLastSave(), options);
+        return this.createWritePromise<ClusterGlideRecord<number>>(
+            createLastSave(),
+            options,
+        ).then((res) => convertClusterGlideRecord(res, true, options?.route));
     }
 
     /**